feat(quality-check): impl api for submit quality

+++ app/src/main/java/com/tebet/mojual/data/models/IOrder.kt
@@ -1,6 +1,6 @@
 package com.tebet.mojual.data.models
 
 interface IOrder {
-    var orderId: Int
+    var orderId: Long
     var orderCode: String
 }
